# frontend/lib/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "allwin API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/matches/{match_id}/odds": {
   "get": {
    "tags": [
     "public"
    ],
    "summary": "Match Odds",
    "description": "free/pro:延迟摘要(每公司每市场最新一条,观察时间 ≥1 小时前);\npremium(odds:history_full):完整快照时间线。",
    "operationId": "match_odds_api_v1_matches__match_id__odds_get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/MatchOddsAvailableDTO"
          },
          {
           "$ref": "#/components/schemas/MatchOddsUnavailableDTO"
          }
         ],
         "title": "Response Match Odds Api V1 Matches  Match Id  Odds Get"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
...
  "/api/v1/admin/xref": {
   "get": {
    "tags": [
     "admin-xref"
    ],
    "summary": "List Xrefs",
    "description": "映射列表。默认返回待处理(needs_review + auto_ok);?status= 精确过滤。",
    "operationId": "list_xrefs_api_v1_admin_xref_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/XrefListResponse"
        }
       }
      }
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/admin/xref/{xref_id}/confirm": {
   "post": {
    "tags": [
     "admin-xref"
    ],
    "summary": "Confirm Xref",
    "operationId": "confirm_xref_api_v1_admin_xref__xref_id__confirm_post",
    "parameters": [
     {
      "name": "xref_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Xref Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/XrefReviewResultDTO"
        }
       }
      }
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/admin/xref/{xref_id}/reject": {
   "post": {
    "tags": [
     "admin-xref"
    ],
    "summary": "Reject Xref",
    "operationId": "reject_xref_api_v1_admin_xref__xref_id__reject_post",
    "parameters": [
     {
      "name": "xref_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Xref Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/XrefReviewResultDTO"
        }
       }
      }
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
...
  "/api/v1/analytics/events": {
   "post": {
    "tags": [
     "analytics"
    ],
    "summary": "Record Event",
    "operationId": "record_event_api_v1_analytics_events_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnalyticsEventBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "已记录/已限流丢弃(无 body)"
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AnalyticsEventBody": {
    "properties": {
     "event": {
      "type": "string",
      "title": "Event"
     },
     "anon_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Anon Id"
     },
     "path": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Path"
     },
     "referrer": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Referrer"
     },
     "meta": {
      "additionalProperties": true,
      "type": "object",
      "title": "Meta"
     }
    },
    "type": "object",
    "required": [
     "event"
    ],
    "title": "AnalyticsEventBody"
   },
   "ApiErrorDTO": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "details"
    ],
    "title": "ApiErrorDTO"
   },
...
   "MatchOddsAvailableDTO": {
    "properties": {
     "match_id": {
      "type": "integer",
      "title": "Match Id"
     },
     "available": {
      "type": "boolean",
      "const": true,
      "title": "Available"
     },
     "tier": {
      "type": "string",
      "enum": [
       "full",
       "delayed_summary"
      ],
      "title": "Tier"
     },
     "home_away_inverted": {
      "type": "boolean",
      "title": "Home Away Inverted"
     },
     "snapshots": {
      "items": {
       "$ref": "#/components/schemas/OddsSnapshotItem"
      },
      "type": "array",
      "title": "Snapshots"
     }
    },
    "type": "object",
    "required": [
     "match_id",
     "available",
     "tier",
     "home_away_inverted",
     "snapshots"
    ],
    "title": "MatchOddsAvailableDTO"
   },
   "MatchOddsUnavailableDTO": {
    "properties": {
     "match_id": {
      "type": "integer",
      "title": "Match Id"
     },
     "available": {
      "type": "boolean",
      "const": false,
      "title": "Available"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "match_id",
     "available",
     "reason"
    ],
    "title": "MatchOddsUnavailableDTO"
   },
...
   "OddsSnapshotItem": {
    "properties": {
     "market": {
      "type": "string",
      "title": "Market"
     },
     "company_id": {
      "type": "string",
      "title": "Company Id"
     },
     "company_name": {
      "type": "string",
      "title": "Company Name"
     },
     "market_phase": {
      "type": "string",
      "title": "Market Phase"
     },
     "source_updated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Updated At"
     },
     "observed_at": {
      "type": "string",
      "title": "Observed At"
     },
     "payload": {
      "additionalProperties": true,
      "type": "object",
      "title": "Payload"
     }
    },
    "type": "object",
    "required": [
     "market",
     "company_id",
     "company_name",
     "market_phase",
     "observed_at",
     "payload"
    ],
    "title": "OddsSnapshotItem"
   },
...
   "XrefItem": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "fotmob_match_id": {
      "type": "integer",
      "title": "Fotmob Match Id"
     },
     "provider": {
      "type": "string",
      "title": "Provider"
     },
     "provider_match_id": {
      "type": "string",
      "title": "Provider Match Id"
     },
     "home_away_inverted": {
      "type": "integer",
      "title": "Home Away Inverted"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "verified": {
      "type": "integer",
      "title": "Verified"
     },
     "method": {
      "type": "string",
      "title": "Method"
     },
     "kickoff_diff_seconds": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kickoff Diff Seconds"
     },
     "review_status": {
      "type": "string",
      "title": "Review Status"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "fotmob_match_id",
     "provider",
     "provider_match_id",
     "home_away_inverted",
     "confidence",
     "verified",
     "method",
     "review_status",
     "created_at",
     "updated_at"
    ],
    "title": "XrefItem"
   },
   "XrefListResponse": {
    "properties": {
     "counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Counts"
     },
     "xrefs": {
      "items": {
       "$ref": "#/components/schemas/XrefItem"
      },
      "type": "array",
      "title": "Xrefs"
     }
    },
    "type": "object",
    "required": [
     "counts",
     "xrefs"
    ],
    "title": "XrefListResponse"
   },
   "XrefReviewResultDTO": {
    "properties": {
     "status": {
      "type": "string",
      "const": "ok",
      "title": "Status"
     },
     "xref_id": {
      "type": "integer",
      "title": "Xref Id"
     },
     "review_status": {
      "type": "string",
      "title": "Review Status"
     }
    },
    "type": "object",
    "required": [
     "status",
     "xref_id",
     "review_status"
    ],
    "title": "XrefReviewResultDTO"
   }
  }
 }
}